requirements.txt
openai>=1.3.0
python-dotenv>=0.19.0
Pillow>=10.0.0
requests>=2.31.0
ipython>=8.0.0
notebook>=7.0.0 